🏗 Heavy Pivot to Yarn Workspaces with React and React native
Expo for Web become a pain to use. The TL;DR is that when trying to
write once for all three platform, I have issues make things work with
how different web and app works. Yes, it is convinient to have on
codebase, but for each line of code there I have to think and compensate
for the massive differences in how web and apps works i.e. Sizing of all
the elements (Mobile doesn't have rem, vw, vh, etc.), Having to use
React-Native Components for Web which is limited, Not knowing whether
this library works with Expo and/or with React-Native-Web, Making
changes to the mobile view destroy the web version, Configuring web
routes without messing the mobile version, No native code at all, and
probably more. Eventhough, my usage are minimal, this issues exist and
has been a massive block for me. At the moment, I am pivotting to yarn
workspaces (trick) so I share most of the logic in the client, but not
entirely the UI so I can differ a bit how they work and compensate for
the platform. Plus, I can use regular react-native if I wanted so I can
add native code as well

- Removed all Expo stuff (for now)
- Added React with CRA and Chakra
- Added yarn workspaces
- Added common package to be shared across clients